# crates/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shelfy API",
    "description": "HTTP API of Shelfy Web. JSON in camelCase, timestamps in unix milliseconds. Errors are `application/problem+json` documents whose `code` is stable; clients map it to their own messages.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1"
  },
  "paths": {
...
    "/api/v1/extension/pair": {
      "post": {
        "tags": ["extension"],
        "summary": "Exchanges a pairing code for the extension's token.",
        "description": "The extension calls it with the code that the web app got from `POST\n/me/tokens/pairing-code`, within 60 seconds, once. No cookie and no CSRF\nheaders are needed, and none is read. 400 `invalid_pairing_code` for a\ncode that is malformed, unknown, used or expired; 422 for a malformed\n`installId`, `label` or `version`; 426 `extension_outdated` for a version\nbelow `minVersion` (the code stays usable); 409 `conflict` when the\naccount holds 50 working tokens. Limit: the sign-in limit, 10 requests a\nminute per client over this route and the `/auth` routes (429).",
        "operationId": "pairExtension",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionPairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Paired: the token, shown once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairedExtension"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [{}]
      }
    },
...
    "/api/v1/me/tokens": {
      "get": {
        "tags": ["account"],
        "summary": "The account's working API tokens (not revoked, not expired), with their\nlast use.",
        "operationId": "listApiTokens",
        "responses": {
          "200": {
            "description": "The tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenList"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": ["account"],
        "summary": "Creates an API token for the extension or the iOS Shortcut; the answer\nis the only time its value is shown.",
        "description": "Needs a sign-in or a re-authentication from the last 5 minutes (403\n`reauth_required` otherwise). 422 for the kind `migrate` (the migration\nCLI signs in with the device flow), for scopes that are not the kind's,\nor for a label over 64 characters; 409 `conflict` when the account holds\n50 working tokens already.",
        "operationId": "createApiToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The new token, with its value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiToken"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/me/tokens/pairing-code": {
      "post": {
        "tags": ["account"],
        "summary": "Creates a code that pairs the browser extension with this account.",
        "description": "The web app hands the code to the extension (`chrome.runtime.sendMessage`),\nwhich exchanges it for its token at `POST /extension/pair` within 60\nseconds, once. Needs a sign-in or a re-authentication from the last 5\nminutes (403 `reauth_required` otherwise). 429 `rate_limited` while the\naccount holds 10 unused codes that have not expired.",
        "operationId": "createPairingCode",
        "responses": {
          "201": {
            "description": "The code and its expiry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCode"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/me/tokens/{id}": {
      "delete": {
        "tags": ["account"],
        "summary": "Revokes one of the account's tokens: it stops working at once. Another\naccount's token, or one that no longer works, is a 404.",
        "operationId": "revokeApiToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The token's id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked."
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiToken": {
        "type": "object",
        "description": "An API token of the account. Its value is never shown again.",
        "required": ["id", "kind", "label", "scopes", "createdAt", "lastUsedAt", "expiresAt"],
        "properties": {
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "Creation time, unix ms."
          },
          "expiresAt": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "When it stops working, unix ms; `null` until revoked."
          },
          "id": {
            "type": "string",
            "description": "Its id (ULID), for `DELETE /me/tokens/{id}`."
          },
          "kind": {
            "$ref": "#/components/schemas/TokenKind",
            "description": "Who holds it."
          },
          "label": {
            "type": ["string", "null"],
            "description": "The user's name for it."
          },
          "lastUsedAt": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "Last request it authenticated, unix ms (recorded at most once a\nminute)."
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenScope"
            },
            "description": "What it may do."
          }
        }
      },
      "ApiTokenList": {
        "type": "object",
        "description": "The account's working API tokens.",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiToken"
            },
            "description": "Newest first."
          }
        }
      },
      "ApiTokenRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/me/tokens`.",
        "required": ["kind"],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TokenKind",
            "description": "Who will hold it: `extension` or `shortcut`."
          },
          "label": {
            "type": "string",
            "description": "A name for the list, up to 64 characters; trimmed, and blank is none."
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenScope"
            },
            "description": "What it may do: some of the kind's scopes; all of them when left out."
          }
        },
        "additionalProperties": false
      },
...
      "CreatedApiToken": {
        "type": "object",
        "description": "A new API token.",
        "required": ["token", "apiToken"],
        "properties": {
          "apiToken": {
            "$ref": "#/components/schemas/ApiToken",
            "description": "The token as the list shows it."
          },
          "token": {
            "type": "string",
            "description": "The value (`shx_…`), for `Authorization: Bearer`. Shown only now:\nthe server keeps its hash."
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "Stable, machine-readable error codes. Each one has a fixed HTTP status.\n\nNew codes may be added within `/api/v1`; existing ones never change meaning.",
        "enum": [
          "bad_request",
          "invalid_cursor",
          "invalid_link",
          "challenge_expired",
          "passkey_invalid",
          "invalid_device_code",
          "invalid_pairing_code",
          "unauthorized",
          "forbidden",
          "csrf_failed",
          "reauth_required",
          "quota_exceeded",
          "not_found",
          "method_not_allowed",
          "conflict",
          "upload_consumed",
          "payload_too_large",
          "unsupported_media_type",
          "validation_failed",
          "provider_key_invalid",
          "capture_blocked",
          "not_available",
          "user_locked",
          "extension_outdated",
          "rate_limited",
          "internal",
          "storage_full",
          "unavailable",
          "timeout"
        ]
      },
...
      "ExtensionPairRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/extension/pair` (contract C2). Fields a later\nextension adds are ignored.",
        "required": ["code", "installId", "version"],
        "properties": {
          "code": {
            "type": "string",
            "description": "The pairing code the web app handed over (43 characters)."
          },
          "installId": {
            "type": "string",
            "description": "A random id the extension keeps while it is installed (16 to 128\nletters, digits, `-` or `_`). Pairing the same installation again\nrevokes the token it held."
          },
          "label": {
            "type": "string",
            "description": "A name for the account's token list, such as \"Chrome on macOS\"; up\nto 64 characters, blank is none."
          },
          "version": {
            "type": "string",
            "description": "The extension's manifest version (`0.2.0`); below `minVersion`, 426."
          }
        }
      },
...
      "FieldError": {
        "type": "object",
        "description": "One field that failed validation.",
        "required": ["field", "reason"],
        "properties": {
          "field": {
            "type": "string",
            "description": "The field, as named in the request (camelCase)."
          },
          "reason": {
            "type": "string",
            "description": "Why the value was refused, for developers."
          }
        }
      },
...
      "PairedExtension": {
        "type": "object",
        "description": "A paired extension's token.",
        "required": ["token", "tokenId", "scopes"],
        "properties": {
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenScope"
            },
            "description": "What it may do: `ingest`, `tasks`, `uploads`, `lookup`."
          },
          "token": {
            "type": "string",
            "description": "The `extension` token (`shx_…`) for `Authorization: Bearer`. Shown\nonly now; keep it where page scripts cannot read it."
          },
          "tokenId": {
            "type": "string",
            "description": "Its id, as the account's token list names it."
          }
        }
      },
      "PairingCode": {
        "type": "object",
        "description": "A code that pairs the browser extension (contract C2).",
        "required": ["code", "expiresAt"],
        "properties": {
          "code": {
            "type": "string",
            "description": "The code (43 characters), for the extension's `POST /extension/pair`.\nHand it to the extension only; it works once."
          },
          "expiresAt": {
            "type": "integer",
            "format": "int64",
            "description": "When it stops working, unix ms: 60 seconds from now."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "An error body: RFC 9457 problem details plus the stable `code`.",
        "required": ["type", "title", "status", "code"],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong. Clients map it to their own message."
          },
          "detail": {
            "type": "string",
            "description": "Developer-facing detail. Never UI prose, user content or secrets."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "The fields that failed validation (`validation_failed`)."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status of the response.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The reason phrase of `status`. Not meant for display."
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`: `code` carries the problem type."
          }
        }
      },
...
      "TokenKind": {
        "type": "string",
        "description": "Who holds a token.",
        "enum": ["extension", "shortcut", "migrate"]
      },
      "TokenScope": {
        "type": "string",
        "description": "What a token may do (§2.9).",
        "enum": ["ingest", "tasks", "uploads", "lookup", "links:create", "migrate"]
      },
...
    },
    "responses": {
      "Problem": {
        "description": "An error, as RFC 9457 problem details with a stable `code`.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "shx_",
        "description": "A scoped API token (`shx_…`) of the extension, the iOS Shortcut or the migration CLI. Tokens cannot call routes that only accept the session."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-shelfy_session",
        "description": "The web app's session, set by signing in. Requests that change state (POST, PUT, PATCH, DELETE) with this cookie must also send `X-Shelfy-Client: web` and an `Origin` equal to the public URL; otherwise they fail with 403 `csrf_failed`. A request with an `Authorization` header is never authenticated by this cookie."
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
...
}